feat(sync): B7 — extract secret-scanner to core/utils standalone module (Phase 1.5)
The handoff calls for secret-scanner to be extractable so
prjct-cloud can reuse it server-side. The file was already
free of path-manager / storage / infrastructure imports — this
commit makes the extraction concrete.

- core/memory/secret-scanner.ts → core/utils/secret-scanner.ts.
  utils/ is the established home for dependency-free helpers;
  the move signals the contract.
- 4 callers updated to import directly from the new path
  (per repo style: NEVER use re-exports — no barrel shim):
  * core/mcp/tools/memory.ts
  * core/commands/capture.ts
  * core/commands/primitives.ts
  * core/services/wiki-ingest.ts
- API unchanged: scanForSecrets() + new SECRET_PATTERN_NAMES
  export for cloud-side telemetry without leaking regex shapes.
- Doc-block warning that future contributors must keep this file
  pure — server-side reuse depends on zero filesystem / SQLite
  imports.

Cloud server vendor strategy: copy core/utils/secret-scanner.ts
verbatim (pure module, zero deps).

Spec: 9a7b071b-7c4e-42a5-ad47-20a652139c86 · B7/8.
