+++ core/services/wiki-ingest.ts
@@ -27,9 +27,9 @@ import fs from 'node:fs/promises'
 import path from 'node:path'
 import configManager from '../infrastructure/config-manager'
 import { MEMORY_TYPES, type MemoryType, projectMemory } from '../memory/project-memory'
-import { scanForSecrets } from '../memory/secret-scanner'
 import { workflowRuleStorage } from '../storage/workflow-rule-storage'
 import type { WorkflowRule } from '../types/storage/extended'
+import { scanForSecrets } from '../utils/secret-scanner'
 import { resolveVaultRoot } from './wiki-migration'
 
 const CAPTURED_SUBDIR = 'captured'
